Initial version of gaussian blur

+++ aos/grayscaleaos.cpp
@@ -87,8 +87,8 @@ void ImageAOS::ToGrayScale() {
     vector<vector<PixelNormalized>> bmpPixelsDataNormalized(bmpPixelsData.size(), vector<PixelNormalized>(bmpPixelsData[0].size()));
 
     //Read all pixels and convert them to grayscale
-    for (int i = 0; i < bmpPixelsData.size(); i++) {
-        for (int j = 0; j < bmpPixelsData[0].size(); j++) {
+    for (long unsigned int i = 0; i < bmpPixelsData.size(); i++) {
+        for (long unsigned int j = 0; j < bmpPixelsData[0].size(); j++) {
             //asign result of  Normalize to bmpPixelsData[i][j]
             bmpPixelsDataNormalized[i][j] = Normalize(bmpPixelsData[i][j]);
             bmpPixelsDataNormalized[i][j] = Intensidad_lineal(bmpPixelsDataNormalized[i][j]);
